Set up Travis CI integration
This is the first version, which uses a hack to adjust TOPDIR
in the Makefile using sed, and then calls `make`. Later, when
a proper `configure` script will be added, there will be no
need to use sed. Also, now this only tests that carat builds
on Linux with gcc and clang. Later, proper tests must be added.