Fix wrong result in gcd_mat()
The variable j was never initialized, and also never reset at the start
of a new matrix row. Hence, gcd_mat() actually only computed the gcd of
*at most* the first row of the given matrix; and possibly did not even
do that, if `j` happened to start out with a value greater than 0.

Also use the correct column count (may differ from the row count).

There even is a chance this might have crashed if j started out with a
huge negative value.