Fix potential crash in vor_vertices
We didn't give enough arguments to voronoi_vertices; prototypes plus
warnings revealed this now.

This could result in a crash, as voronoi_vertices writes into the
last pointers given to it. We don't care what it writes in there, so
we just give it a pointer to something we don't need; in this case
the variable j, twice.