Refactor scpr to satisfy clang static analyzer
Untangle the cases with a single input file, and multiple input files.
They were handled "together" but actually shared little code. This complicated
things enough to confuse the static analyzer to believe that there might be
access to uninitialized data.

By explicitly separating the two cases, we avoid the warnings, and also make
the code perhaps a bit easier to follow through.