Fix allocation sizes, reported by clang static analyzer
The TYPE sizeof(TYPE) in the size computation for malloc/realloc calls
did not fit the actual return type. In most cases, this luckily resulted
in no bugs and at worst some overallocation.

But in a few places, we used `sizeof(int)` instead of `sizeof(int *)`, and
on 64 bit systems, that could result in a too small allocation.