Make code C++ compatible
Being able to compile the code with a C++ compiler helped me track down
several further issues. Also, in the future it might be interesting
to convert some or all of the code and use some basic C++ features to
greatly simplify the carat code, and in particular the memory management
in it.